refactor: move validate-mocha into recorder/mocha
This functionality is no longer called from `lib` and does not deserve a
dedicated component.
