--- components/validate-mocha/default/.env
@@ -1,3 +1,0 @@
-node
-test
-browser
